Move RunAsync to backend
