+++ bottles/frontend/utils/meson.build
@@ -5,7 +5,6 @@ bottles_sources = [
   '__init__.py',
   'connection.py',
   'gtk.py',
-  'threading.py',
   'common.py',
   'filters.py',
 ]
