+++ bottles/frontend/views/bottle_versioning.py
@@ -21,7 +21,7 @@ from gi.repository import Gtk, GLib, Adw
 
 from bottles.backend.models.result import Result
 
-from bottles.frontend.utils.threading import RunAsync
+from bottles.backend.utils.threading import RunAsync
 from bottles.frontend.utils.common import open_doc_url
 from bottles.frontend.widgets.state import StateEntry
