+++ bottles/frontend/views/list.py
@@ -15,15 +15,13 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 
-import logging
 from datetime import datetime
 from gettext import gettext as _
 from gi.repository import Gtk, GLib, Adw
 
 from bottles.backend.models.config import BottleConfig
 
-from bottles.frontend.utils.threading import RunAsync
-from bottles.backend.runner import Runner
+from bottles.backend.utils.threading import RunAsync
 from bottles.backend.wine.executor import WineExecutor
 from bottles.frontend.utils.filters import add_executable_filters, add_all_filters
