+++ bottles/frontend/views/new.py
@@ -16,17 +16,12 @@
 #
 
 import os
-import re
 from gettext import gettext as _
 from gi.repository import Gtk, Adw
 
 from bottles.backend.models.config import BottleConfig
-from bottles.frontend.utils.threading import RunAsync
-from bottles.frontend.utils.gtk import GtkUtils
+from bottles.backend.utils.threading import RunAsync
 from bottles.frontend.utils.filters import add_yaml_filters, add_all_filters
-
-from bottles.backend.runner import Runner
-from bottles.backend.wine.executor import WineExecutor
 
 
 @Gtk.Template(resource_path='/com/usebottles/bottles/new.ui')
