+++ bottles/frontend/widgets/importer.py
@@ -17,7 +17,7 @@
 
 from gi.repository import Gtk, Adw
 
-from bottles.frontend.utils.threading import RunAsync
+from bottles.backend.utils.threading import RunAsync
 from bottles.backend.utils.manager import ManagerUtils
