+++ bottles/frontend/windows/display.py
@@ -15,12 +15,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import os
-from gi.repository import Gtk, GLib, Adw, Gdk
-from bottles.frontend.utils.threading import RunAsync
+from gi.repository import Gtk, GLib, Adw
+from bottles.backend.utils.threading import RunAsync
 from bottles.backend.logger import Logger
-from bottles.backend.managers.runtime import RuntimeManager
-from bottles.backend.utils.manager import ManagerUtils
 from bottles.backend.wine.reg import Reg
 from bottles.backend.wine.regkeys import RegKeys
